Add files via upload
Avoid to hardcoded IDs which may not exist, for create a access for a specific client, the expectation is some code it can receive the name of the client and assign/grant the access, in this case it'll only work for Maria Stefania, although an user example must be provider, shouldn't be hardcoded. Please made the udpates and merge.